Add Masks to the dxNumberBox
